% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cd.R
\name{export_report_cd_patti}
\alias{export_report_cd_patti}
\title{Esporta report CD in xlsx}
\usage{
export_report_cd_patti(report, titolo = NULL, focus = NULL)
}
\arguments{
\item{report}{dataframe da report_cd_regioni().}

\item{titolo}{titolo interno al report (chr).}

\item{focus}{nome per file.}

\item{ambito}{...}

\item{gruppo}{...}

\item{...}{...}
}
\value{
Un file xlsx con apertura per patto e fase procedurale.
}
\description{
Esporta report CD in xlsx da template "stato_template_regioni.xlsx"
}
