% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cd.R
\name{report_cd_regioni}
\alias{report_cd_regioni}
\title{Esporta report in formato CD}
\usage{
report_cd_regioni(
  perimetro,
  ciclo = NULL,
  ambito = NULL,
  gruppo = NULL,
  po_ls = NULL,
  focus = "elab",
  export = FALSE
)
}
\arguments{
\item{perimetro}{Dataset di classe perimetro.}

\item{ciclo}{in formato x_CICLO.}

\item{ambito}{in formato x_AMBITO.}

\item{gruppo}{in formato x_GRUPPO.}

\item{po_ls}{vettore di PO in formato x_PROGRAMMA.}

\item{focus}{nome per file.}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per regione e fase procedurale.
}
\description{
Report con apertura per regione e fase procedurale rispetto al focus selezionato
}
